requirements.txt
# OpenClaw MeshCore - Dependencies
# Install: pip install -r requirements.txt

# MeshCore Python SDK (async serial/BLE/TCP communication)
meshcore>=0.1.0

# HTTP client for AI APIs (Ollama + Anthropic)
httpx>=0.27.0

# YAML configuration
pyyaml>=6.0

# Optional: GPIO control (install on Raspberry Pi / Linux SBC)
# RPi.GPIO>=0.7.0
# lgpio>=0.2.0
